+++ api/TimeTracker.Application/Activities/Commands/UpdateActivity/UpdateActivityCommand.cs
@@ -7,8 +7,8 @@ namespace TimeTracker.Application.Activities.Commands.UpdateActivity
     {
         public Guid Id { get; set; }
         public string Description { get; set; }
-        public int StartInMilliseconds { get; set; }
-        public int EndInMilliseconds { get; set;}
+        public long StartInMilliseconds { get; set; }
+        public long EndInMilliseconds { get; set;}
         public Guid? ProjectId { get; set; }
         public List<Guid> TagIds { get; set; }
     }
