+++ api/TimeTracker.Domain/Activity.cs
@@ -5,7 +5,7 @@
         public string Description { get; set; } = string.Empty;
         public List<Tag> Tags { get; set; }
         public Project? Project { get; set; }
-        public int StartInMilliseconds { get; set; }
-        public int? EndInMilliseconds { get; set;}
+        public long StartInMilliseconds { get; set; }
+        public long? EndInMilliseconds { get; set;}
     }
 }
