+++ api/src/Models/CreateActivityDto.cs
@@ -10,8 +10,8 @@ namespace TimeTrackerApi.Models
         public string Description { get; set; }
         public List<Guid> TagIds { get; set; }
         public Guid? ProjectId { get; set; }
-        public int StartInMilliseconds { get; set; }
-        public int? EndInMilliseconds { get; set; }
+        public long StartInMilliseconds { get; set; }
+        public long? EndInMilliseconds { get; set; }
 
         public void Mapping(Profile profile)
         {
