+++ api/src/Models/UpdateActivityDto.cs
@@ -11,8 +11,8 @@ namespace TimeTrackerApi.Models
         public string Description { get; set; }
         public List<Guid> Tags { get; set; }
         public Guid? Project { get; set; }
-        public int StartInMilliseconds { get; set; }
-        public int? EndInMilliseconds { get; set; }
+        public long StartInMilliseconds { get; set; }
+        public long? EndInMilliseconds { get; set; }
 
         public void Mapping(Profile profile)
         {
